use cases

+++ src/Backend/Infrastructure/Migrations/Versions/Version0000002.cs
@@ -25,7 +25,7 @@ public class Version0000002 : Migration
             .WithColumn("InvestmentFixedType").AsInt32().NotNullable()
             .WithColumn("Profitability").AsString(100).NotNullable()
             .WithColumn("MaturityDate").AsDate().NotNullable()
-            .WithColumn("IR").AsInt16().NotNullable()
+            .WithColumn("IR").AsInt32().NotNullable()
             .WithColumn("UserId").AsInt64().NotNullable().ForeignKey("FK_VariableIncome_User_Id", "Users", "Id");
     }
